content/base/test/test_bug503473.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=503473
-->
<head>
  <title>Test for Bug 503473</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=503473">Mozilla Bug 503473</a>
<p id="display"></p>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 503473 **/

netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
var prefs = Components.classes["@mozilla.org/preferences-service;1"]
          .getService(Components.interfaces.nsIPrefBranch);
var gOriginalHtml5Pref = prefs.getBoolPref("html5.parser.enable");
prefs.setBoolPref("html5.parser.enable", true);

SimpleTest.waitForExplicitFinish();

function done() {
  var iframe = document.getElementById("iframe");
  var divs = iframe.contentWindow.document.getElementsByTagName("div").length;
  is(divs, 0, "Div wasn't blown away.")

  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
  prefs.setBoolPref("html5.parser.enable", gOriginalHtml5Pref);

  SimpleTest.finish();
}

</script>
</pre>
<div id="content" style="display: none">
  <iframe id='iframe' src="file_bug503473-frame.sjs">
  </iframe>
</div>
</body>
</html>
